(shmif) reslot NEWSEGMENT reqid
This has apparently been confused for a long while, and survived simply because
nothing really enforced the matching between the SEGID supplied in the SEGREQ
versus the response that was actually received. The core issue was that the
[0].iv slot is always reserved for the file descriptor in events that carry a
reference handle (such as a NEWSEGMENT), but the other layers used this slot to
carry the request id cookie.

FossilOrigin-Name: 58eb9457f6fd066ffbd281f1d0fa90f2cbfd443d0f1ce75d1b4e62047851cd54