(tui) add skeleton for screen store/restore
Halfway mark for two related features. One is to be able to take a screen part
of a tui connection and serialise into a line format. This is a prerequisite
for the remote protocol part (being able to let another process, say an arcan
instance, take the job of rendering). This should also be able to handle some
of the visual anomalies from drag resize when the terminal is slow and old
contents gets evicted to the scrollback after resize.

The second part is the user initiated ability to clone the contents of the
current screen, spawn it into a new thread or process and have it act as a
passive 'screenshot' or clipboard buffer (clipboard entries gets added to the
screen and scrollback of the copy window).

FossilOrigin-Name: 7ffc59b91d60c657a2a0029d15612a68dbecd1b784e30ee69b9e60d0b342b7e1