(build) expose installation overrides
The old way of setting install destinations was an ugly patchwork from
the days of the W32 port and some paths were being set in broken ways,
expose the variables so that packaging configurations can override them.

In addtion, the BUILDTAG is now force-set if the git- checkout probe
doesn't work in order for git archive- derived builds to work without
explicit DISTRTAG setting.