(platform/posix) plug leak on named socket fail
This fixes a subtle edge case that could result in a descriptor leak and
semaphore leak when an external listening point is allocated and the
name is already used. This becomes and issue when/if the running scripts
periodically tries to reopen the same connection point, which durden
could be provoked to do in the case of an older implementation of
terminal groups.