(core/lua) draw tesselation-surface 2-side
Two fixes to tesselation surfaces, one is being a null-deref due
to a missing check on the invariant of going from a tesselated
surface to a normal one.

The other is to enable the tesselation-surfaces to be drawn 2-sided
as quite a few possible 2d/pseudo-3d effects relies on that.