(waybridge) seccomp filtering etc.
a few things in this one,

First is a commandline toggle for deferred buffer releases and
buffer release on client rebuild in recover mode.

Then two missing messages forwarding transform and scale info.

Last is the first list of seccomp syscalls to permit if the
bridge was started with -sandbox.