(core/lua/platform) improve GUID handling
This adds chacha20 as stream cipher and csprng and reworks all
use to go via a platform- arcan_random api (simplified arc4random).

GUIDs are generated using the rng and we now support notifying
target_devicehint style fallback connection point about the current
accepted GUID.

This is primarily to handle the case with auth- connections that
don't go through a REGISTER state, meaning that there is never any
GUID for the appl to use for tracking between launches.