(shmif) enforce migrate
This behavior is somewhat up for evaluation. Previously, the decision
was to only forward device hints that were for suggested direct
migration, but to enable this would require a lot of development work.

Changing it to default-migrate, forward on failure behavior so that
the client still get the device-node hint if the migration could not
be performed from the context of shmif, otherwise treat it as a normal
state-3 reset.

FossilOrigin-Name: 347d38e02d0e129eca04814249d519a60376e6369744a4616ce68232b13b511e