(platform) device- open management improvements
Extended the preinit state with an _open wrapper that wraps the
device open calls to pre-mastered descriptors. This should allow
the binary to work under egl-dri as suid safely, but also prepare
to allow a subprotocol and descriptor-passing to deal with device
management on setups where normal uid/gid isn't enough.