(3dbase) reindex build-box
The build-box was only ever used as debug geometry, so texture
coordinates and winding order was not correct. Should extend and
test this as a helper for %6==0 nmaps as well so that it gets
treated as single logical entity but rendered with different
texture sets for easier skyboxes.

FossilOrigin-Name: 0da2c0f2446428e4403e3afcf17c79b6688e13539f3d94dfda48f45e987d6518