(build) externalproject should use external/src directly
previously, before we opted to use an internal fork of openAL, we
pulled from their git. When migrating away from that due to all the
upstream breaks, the git approach was kept but rereferenced to
clone locally. This is unnecessary and can be circumvented by pointing
the source directory directly without a download step.