(core+) first round in adding 'conductor'
This is the first of many steps towards a sane synchronization
subsystem.

The current synchronization approach survived for about 15 years, but
scales terribly and is not fit to deal with the reality of current and
coming displays, and a number of other caveats.

The solution is to move the main processing loop into what will
be referred to as the 'conductor' where all dynamic / asynchronous
operations will be registered and tracked.

The plan for how this is going to go down is a bit sketchy still,
and this round only moves the setup from _main into _conductor.

The next steps will be (rougly):

1. add support for thread + futex- managed resize operations and
audio buffer transfers to cut down on latency.

2. add support for fenced buffer uploads and downloads.

3. add support for 'focus' target, a single frameserver connection that
gets synch- bias, using previous draw-cost estimates to try and align
display synch to falling edge based on the focus target.

4. rework platform video synch so that it can indicate when a
gpu is locked and provide a fence for releasing.

5. rework platform video synch so that it registers its output
timings and behavior for each active display.

6. log output system that permits graphing as this will be hard to
verify and monitor.

7. move synchronization strategy settings etc. out from the normal
platform and into the conductor.