(core+lua) initial support for sliced stores
This is part of a much needed cleanup of the 3D subsystem as part
of preparation for a larger VR demo. In this push, all the wiring
is added to in order to support cubemaps and 3dtextures. This is
done by:

1. updating build_3dbox to allow a per-face submesh split.
2. adding sliced storage support to agp/video layers.
3. adding image_storage_slice function to lua layer.

The image_storage_slice function takes a destination vstore, an
expressed slice type and a set of slice sources. These sources
are then used to populate the GPU side of the destination vstore
so that it covers either a cubemap or a 3d-texture with the
contents of the slice-set.

FossilOrigin-Name: a055961ba65698625f8bd069827a049f9bd4f061a055eb1feb58b240eb11e376