(cure/lua) rendertarget id to shader uniform
This add support for the lua function rendertarget_id that can be used
to arbitrarily tag rendertargets in a way that will be communicated to
the shaders that are used. This is mostly for edge cases where multiple
renderpasses work on the same pipeline but you want different behaviors
without geneerating n new shaders.