(platform/shmif) restructure to allow modifiers
As part of supporting DRI3.1/zwp_dma_buf some of the platform layer
and the device- push in the corresponding shmif event need to be
reworked. This is a big task that will need to be staged over many
commits and with some time in between.

The basic checklists hides in egl-dri/video.c, but basically the
queried metadata need to be copied in the cardhandle_ function,
shmifext need a flag to treat as multiplanar and the corresponding
handle- transfer event also need to mark for multiplanar support.

After that, the platform/frameserver part needs to be able to map
the handles with the correct extension and tag for each plane
(EGL_DMA_BUF_PLANE0_MODIFIER_LO_EXT, ...) and the EGLImage conversion
step needs to be able to switch format and mapping.

Lastly, the shader management need to be able to treat this as a
multistore frameset, which blocks certain parts of the Lua API unless
we accept the cost of a FBO+reblitting stage with a conversion shader,
which - given the case where these are useful - is too expensive.

With that round of refactoring over, adding egl-external platform
with streams/alloc support for client side should be relatively
easy.

FossilOrigin-Name: a9aa9af0ffb8c184267094a307996027958c4862d5a7a4a2673c73a355c63ea7