(platform/egl-dri) fake-set dpms on unmap
This is something of a workaround, yet close to the real solution.  When
a display gets mapped to an invalid ID, we should mark the display as
unmapped, though this is a rare event in the display management
lifecycle and has only recently appeared when experimenting with VR.

The normal path is that displays gets unplugged and plugged or when VT
switching is activated - both paths work for that. This case needs a
hybrid, but without triggering any events to the higher layers. Due to
the pending conductor- rework and related atomic-modeset changes, the
DPMS state is only fake set as the way it is done currently is not
guaranteed livelock safe (needs to be deferred so it isn't triggered in
an unsafe state).

FossilOrigin-Name: fb1a4becaac19e5178d571717bb61db3dc1349ef069951b6014f20d405a886bc