(term/tui) add exec_state callback
This push adds state- forwarding (for exit, pause and unpause) to
the tui client, along with the code in the terminal to signal the
group accordingly.

It also adds back some light latency-vs-usage tradeoff code to the
heuristics in the terminal as the last change (neverending story)
failed to deliver as well.