(shmif) SEGID_DEBUG default implementation hook
This relates to the previous "dynamic loading of TUI" in b3cd666. Normally, the
DEBUG subsegment comes from an explicit push from the policy layer to the
client as a means of requesting more data. Not many (= none) actually maps and
uses this for anything meaningful, and just lets its go to the default
silent-drop state. This patch prepares for an alternate path where there is a
default implementation which redirects STDERR to a dynamically loaded TUI
session and provides inputs for easily preparing / attaching a debugger to the
source process.