(shmif) use max_align_t for alignment calculations
We erroneously used uintptr_t for alignment calculations in apad
region to deal with proper alignment for atomics, which might not
be enough on certain architectures (arm7 comes to mind), thus
causing a SIGBUS for alignment in rare edge cases (gamma controller
and vr bridge to be exact). This calculation enforces the apad to
be strict 16- byte aligned for start, and other members to match
max_align_t. This might be a no-op on x86 etc. but may also need
a full recompile of the shmif ecosystem. Good thing a release is
pending soon.