(platform/evdev) privsep + hotplug rework
This push adds support for GPU connector hotplug events, and reworks
much of the privilege separation to be more flexible in order to reuse
much of the same code between linux and openbsd. Much more testing needs
to be done, but results look promising on the first set of devices.

Also fixes regression where the option of adding an override config
directory to inotify scoped the setup incorrectly so the inotify events
never fired, thus breaking evdev hotplug. Fixed, and also added a
workaround to broken inotify by always permitting the rescan action.