(build) temporary openBSD build fixes
Due to the changes to platform_device_open with the linux changes,
the corresponding openBSD code should be reworked. Pending that
work, the temporary measure is to revert to normal open/... use.

This will likely break native use on OpenBSD, roll back to latest
tagged version if that is a problem, the real version is coming
soon.