(arcterm) debug subwnd sorted, more fields added
Turned out to be a usage doh that should be better covered in the TUI
API. Having a process->update->refresh cycle as a function that can be
triggered in the event handler is a big no no for two reasons. One is
the possible recursion, and the second is that the resized event handler
will be invoked before the buffers etc. have been properly reworked
(this might be a behavior to change and protect against) regardless of
any other outcomes.