(tui+) change semantics for subwindow allocation
Previously, the client was required to map and destroy subsegments
that was requested but not handled. Now the client has to indicate
that the segment was consumed, or a default hidden implementation
might be activated.

FossilOrigin-Name: c64ee269ad4023041e6b573bdd688883f8092ed65f7c4c5ed7fca25652cfc6e0