(platform/posix) reworked psep+tty switch
This is a big one with many nuances that I doubt got right this time
around (either). With the psep- switch and running from a script we may
run into the problem that we don't have permission to perform the tty-
switching dance in the lower privileged process. This promted that all
the related code gets moved into the psep part and we extend the socket
protocol to carry data about tty switching requests and so on.

FossilOrigin-Name: c3f8e66241efc9959baaef9f2d93c107fc6cb33f99a44541931d2f560a9913c3