(tui) writestr helper + erase_region "fix"
Some code scenarios led the developer down the road of passing a normal C
string rather than byte array + length field, forcing the strlen+casting
pattern. To make it more clear when and where this is going on, a helper
function (writestr) was added that performs this conversion.

An 'easy' pattern for updating the screen tui wise has been to just start a new
frame by erasing all the cells of the previous one. This is bad in the sense
that a full raster+transfer operation is also needed.

To make it slightly less terrible when the developer doesn't update only the
dirty regions, the erase_region (indirectly, erase_screen) calls now checks the
cells against the default attribute, and only flags the cell as updated when
they differ.

FossilOrigin-Name: 9c6b997e453c17581c128a5bc9df831d37469dceb669ace518934533774d419e