(arcterm) more timeout modifications
This is likely to break on either OSX or one of the BSDs by live-
locking or eating too much CPU - but the timing and presets are better
on linux than the current setup. If it is a problem on the other
platforms, add special ifdefs there (it pertains to polling ability on
ptys).