(lua/core) add depth buffer control to tesselation
For certain pseudo-3D effects, image tesselation can be used to convert
an image from being treated as a normal 'quad' like surface to one with
a higher number of vertices. This also adds the cost of switching the
processing pipeline back and forth between 2D and 3D processing, which
can be costly (depending on platform). To ease this a bit, this patch
adds controls for specifying that vertices won't offset in the Z axis,
meaning that the depth buffer isn't needed.