(core/lua) dual-expose constants, :self fix
When mapping the Lua VM, the constant exposure function was erroneously
called twice. This, in the case of crash recovery, led to the actual
crash message being zeroed out before the scripts get a chance to read
it.

In addition, the man- refers to :self as a way to define the fallback
appl without having to set the appl- name explicit twice, but the
translation was never applied.

FossilOrigin-Name: e23649d0de6c2b4b2a2909fb1c40d5aedcaa71f145e0661800eeb56670bf39b3