(core/lua) extended resample_image
A long held problem has been that there was no 'easy' way to snapshot
the current state of a frameserver for composition like effects. The
workaround had been to define rendertarget and an intermediate surface,
share storages, forceupdate, snapshot, etc. Since this process is
already implicitly present in 'resample_image' - this function was
extended to allow a destination storage. This allows for blurring
effects etc. that quickly snapshots the state of any store. A thing to
think of for the future with handle-backed stores is that the shader
need to take the format variant into account with this shader, and the
mechanism for that is yet to be determined.

FossilOrigin-Name: 6d6cf5c9cf4914b385be4631e77818c1441bf6c7bda9ef32a854c06b45ad03ef