(core/lua) add system-scripts namespace
This is a big change that might have some special side-effects, though
the strictness of system_load should keep things in check. Basically it
adds a final namespace for system_load to search when looking for
scripts.

Previously, the order was APPL(_TEMP) -> SHARED_RESOURCE. Now it is
APPL(_TEMP) ->SHARED_RESOURCE -> SYSTEM_SCRIPTS. The intention is for
distribution / global packaging implementations of certain keyscripts
that keep on getting copied and re-used, typically mouse.lua and
symtable.lua (now renamed to keyboard.lua) - and slowly refactor and
cleanup some of the many lessons learned from durden so that they can be
reused.

To make collisions less accidental (not that they would have much impact
given the search order), these scripts are kept in a builtin
subdirectory.