(core/lua) change return of reset_image_transform
The previous return value, i.e. the total count of pending transforms
was wholly useless. Since nothing would / could get any use of that,
it has been changed to the number of ticks left in each transform slot.

This fills the case where you have a pending animation (single), and an
event occurs that would prompt the animation to be cancelled and swapped
for another (think cross-fade between two objects) but you want the
real-world time elapsed for the initial animation to match.