(egl-dri) reworked EGL context management
This should fix the issues introduced with (primarily) AMD GPUs in
recent version of MESA. The ugly details are somewhat hidden deep inside
the stack, but relates to the fact that the config received from
eglChooseConfig doesn't necessarily match the buffer format that the
related GBM buffer may want.

The refactor does a number of things that need testing for all the
permutations of single/multi-monitor, VT switch, suspend/resume,
plug/unplug, EGL streams and GBM based and mode switching - so it might
be a rough ride for a while.

The big change is that the EGL structures for Context and Config is
now handled per output display, and they are then shared with an outer
surfaceless context in which the GPU-local data is handled. This should
hopefully have the effect that we can support outputs with different
formats (10-bit + 8-bit + 565- for instance).