(core/lua) add script-controlled resize-block
In some odd situations, the script might want/need to block/defer a
resize operation until some animation has finished. The normal state
is that when the resized event is triggered, the backing store has
already been updated.

With this change, a target_flag can be set to require a stepframe in
order to acknowledge a new resize. This provides ample room for the
scripts to use resample_image to obtain a copy of the buffer contents
before the resize was activated.

The "obvious" target for this is to keep the original during a resize
animation and release / blend when the animation has finished.

FossilOrigin-Name: 01b4682994478f980d4c018df5872d573d63242113fe56afb8272d3c0c1b63db