(lua) improve error reporting in debug for cberr
The adds an error function (and cleans up the wraperr stuff) to the
pcall that, in debug builds, map to adding a debug.traceback() into
the msg before activating fallback etc.

The replacement, lua_acall also adds a safeguard against an ugly form of
asynch- race, it was a noted possibility with the direct-to drain
enqueue method - but essentially - with too much backpressure in the
event queue, (say from spawning a few hundred video decode instances)
chances would be that something critical, like a terminated- event would
jump ahead of the queue, unref:ing the otag needed to find the callback
in lua space. This case was covered for with the ugly event queue
rewrite call where fsrv- event otags were being erased after the fact
but this isn't being applied in the case where the backing object is
still alive.