(core/shmif) handover convenience functions added
This extends the interop support with helper functions to deal with
spawning a new process that inherits and builds on a HANDOVER
subsegment. This should cover one of the major prerequisites for dealing
with chains of terminal-less TUI clients.

Also adds (mostly hollow) hooks for advanced dirty regions, (walking
chunks in buffer) and (hopefully) per-scanline chase- or tail- the beam.
This also comes with hooks for 'next-deadline' estimate so that clients
can better time their polling and updating.