(lua) report api misuse to next recipient
The detailed but costly error reporting gets set as the forced default
for the time being, will evaluate at a later stage but it seems to do
more good than harm.

Also add the arcan_fatal API misuse trigger to the 'last_crash_source'
propagation so that not only "normal" scripting errors gets reported
but also out-of-spec API use.