(lua) acall on nested callback stack offset
The previous rework of acall was only partially correct as the safety
check for the function at the top of the stack wouldn't work in cases
where there was a more complicated stack to work with. This broke the
functions that act like calctarget/access_storage/image_tesselation.

Fixed by checking the function position based on the set number of
input arguments.