(core+) let process title reflect state
This doesn't apply on linux (no setproctitle and the argv[] write trick
just isn't worth the ugly) - but for the others we start setting more
descriptive names of the various subprocesses.

The first step is just making sure that the arcan instance has a proc-
title that reflects the name of the appl that is running, and that the
psep- manager indicates its purpose. Later we will do the same when we
have an execution delegate and privsep the parsers.