(freebsd) update to use psep_open
FreeBSD was forgotten during the whole 'refactor into suid-privsep'
on egl-dri video platform and was thus quite painful to get to run.
This should be fixed and sans some input/repeat rate weirdness due
to the way the tty is set up - should be working.

The input issue will likely be deferred until FreeBSD 12.0 as they
are adding some kind of evdev backend support, so maybe we just
switch to that.