(shmif) switch egl-display lookup function
Some new issues have been introduced with recent versions of mesa that
makes the shmifext_setup fail in contexts like setting up the proxy
wayland-egl display. This does not directly fix the situation and it
ties in with the whole external-platform thing, but at least suppresses
some mesa related errors.