(waybridge) slight rework of frame-cb management
This does not fix the bad subsurface bug that has been haunting for
a while (i.e. gedit -> hover on button -> client gives up on motion
outside subsurface. It is some kind of weird asynch race tied to
both buffer releases and frame callbacks, and the 'defer-release'
option alters the behavior. Won't waste more time on it this push,
something for 056.