(posix) psep_open fallback to O_RDONLY
This prevented backlight to work for a number of devices. Before,
the privileged part of psep_open failed (in spite of whatever mode
the client requested) if the device could not be opened RDWR. This
would fail for a number of sysfs nodes, getting the client to give
up on persuing the path further. Just attempt RDONLY on failing to
get the node RDWR.