(shmif/lua) add data model for eyetracker
This plugs the engine/lua/shmif side for supporting an eye tracker
being used for input. Though the driver side is spotty on linux and
a lot of rev. work to get a usable one, a compromise for now is to
pop a windows WM and simply stream the samples to a local/external
shmif input driver.