(core) switch arcan to _lwa if ARCAN_CONNPATH
This change checks the ARCAN_CONNPATH environment in the normal
arcan binary. If it is detected, it simply switches to arcan_lwa
with the same incoming arguments.

This works around the possible resource conflict in that spawning
the normal arcan process from a context where there already is one
running (which should be the case if ARCAN_CONNPATH is sit).

FossilOrigin-Name: 6459929bb6be1b40783f9306ce4746fd859eb24eaf9d3e7b7b69d15362b21dd9