(core/shmif/sdl/build) harmonize shmifext- backend
A few subtle things happen here which are not without their problems.
With this modification, Linux builds on the SDL VIDEO_PLATFORM will
share the same shmif- backend as the egl-dri one. This means that a
build of egl-dri platform and sdl platform can co-exist, making it
possible to have an arcan-egl and an arcan-sdl binary at the same
time.

The downside is that clients using OpenGL or Vulkan will, for the
arcan-sdl server, fail to transfer buffers in an accelerated way.
The APIs to do this are technically there, but the driver support
is not, but it seems to be coming (EXT_memory_object_fd).

It requires more testing, and there are a bug or two, particularly
something with the FBO based indirection that makes glCube in
(tests/frameservers) work, but not arcan_lwa.

FossilOrigin-Name: f4950e18dceb85924f2e616f714cc8595b19f56e4fd890edc3a4662e729e58fc