(core++) conductor synch. consolidation
This was a big cleanup item that had been pending for a while. Major
changes to the video platform layer to better deal with different
synchronisation tradeoffs. The previous strategies have been scrapped
in favor of a shared set exposed by the conductor.

While far from complete (and lots of heuristic tuning, analysis tools
etc. needed to be written), this should represent the big haul towards
useful VFR support, better / real parallelism and so on.

The lua layer was also extended with a function to specify the
focus_target, that is some frameserver that should be prioritised if the
current strategy supports such a thing.