(shmif/tui/lua) add extended subsegment request format
An oversight from the previous API was the lack of any option to hint
the desired size and relative positioning at the TUI level, though
SHMIF supported it for splits.

With this change, a tui client can specify either a split direction
or position bias direction along with contents hints to provide better
WM integration.