(core/headless) conductor/synch adjustment
Changed the responsibility for 'fake' vsynch behavior to also be a part
of the conductor if triggered from the platform layer. The headless
platform is the first out to use this, possibly that the SDL ones might
as well further down the road.

With this the headless platform needs an input source and the option to
use an external display (encode frameserver) and it should be an
interesting option for some use cases.