(lua) iterate- clock_pulse
This is an unfortunate (yet likely permanent) workaround due to the poor
adherance on lua- script side as to the arguments of clock_pulse.

The initial idea was that the scripts working on clock_pulse should be
able to support multiple ticks in one batch as to minimize the number of
C to VM transitions in the event of multiple ticks stacking up.

With this patch we, instead, invoke a clock_pulse(GLOBAL, 1) for each
tick that has been batched. Should the old optimization turn out to be
useful, consider re-adding as a separate entry point, i.e.
clock_pulse_batch to help conforming scripts.

FossilOrigin-Name: 2f964e0e6351c82d23929f31e2766ffbeff9c00cea1ffd203e33c6c1956ab9f4