(encode) input- fixes to vnc mode, png mode added
This should fix the mouse cursor mapping and the keyboard button
translation for the vnc server stage of encode. Also add a png-stream
mode for storing a specific range of frames into png. This helps with
validation of encode source, primarily for the headless video platform
as a better 'full-stack' approach to automated testing.