(headless) safe-guard against missing -w -h
The (old) CLI approach of setting initial dimensions have usually
played out is part as the display itself tend to know what its starting
default it. This doesn't work for the headless platform, resulting in
hard to detect '0-size' FBOs being created. Adding a fatal-check for
that state and provide defaults for headless.