(waybridge) squish-import xwl support
This adds support for running XWayland and bridging one or many X
clients over one or many XWayland/arcan-wayland instances. The
implementation is young still, the biggest blockers being popup/
grab support. The slightly neat thing about this implementation is
that if we chose to extend xarcan with a rootless mode, the heavy
part, i.e. the pseudo-WM, can be reused.