(egl-dri) nvidia-blob fixes
A few things to note with this, the setup is rather slow still but
that seems to be driver issues. The breakages from before were a.
one case of 'derp' on my behalf where crtc_id and crtc_index were
mistreated.

The more interesting is the AUTO_ACQUIRE parts, if it's false we
seem to stall on eglSwapBuffers indefinitely.

The vsynch- acquire stuff seems to fail still, forcing the fallback
to CLOCK. This might be because of ATOMIC, though the legacy path
didn't yield better behaviour either.

Now, the weirdest thing of all is WHY scanout worked previous to
this, and I don't really have an explanation. So we try GBM first
and then fallback to try streams if that fails to deliver. But even
with nouveau erased from the system and only blobs, not even drm
noveau drivers, it still managed to not only get a GBM path but
actually render.

FossilOrigin-Name: f3fbd69438aeecc457a98123349e21251cc906831aaec9700a7112fba8689d73