(engine+) rework agp_rendertarget_swap
This is still somewhat incomplete and will likely result in bad
output for shmif_ext contexts and wrong resize behavior for lwa
but needed this as a checkpoint between multiple machines. What
is changed is that the rendertarget_swap operation now maintain
buffering to counter the long-standing 'glitch on first+resize'
bug. This highlighted another issue with first-frame hints that
does not seem to be applied properly (origo_ll).