(agp/shmif-egl) rework rendertarget_swap
This was a long and involved refactoring which is guaranteed to leave
some issues to be solved. Simplified the rendertarget management to
be able to mutate from single- to swap-chain style attachments.

This is to open up for a streams- based client provider, and to allow
rendertarget color attachments to be scanned out directly.

FossilOrigin-Name: 04e96a92ee7b47d6c3b0cce2cd6ec9c540e965beece6ab9b9107bd17b9bcfc17