(tui) dropped screen- switching
This was the feature of a tui context to have multiple screens as
a convience of switching between different batches of state.

The feature was poorly planned, confusing and never actually used.
Its substitute will likely be to forward the option of saving and
loading state as an opaque buffer line format. This code exists
already for the copywnd feature, so not a big deal.