(platform/agp) change default clearcol, rtgt-proxy
This changes the default rendertarget clearcolor to be slightly grey
so that they can be distinguished from a failure-state more easily.
Also adds a rendertarget_proxy option for the platform layer. This will
be covered in future patches, but it is primarily intended for a path
where we have a rendertarget mapped to an output, but are not able to
scanout it directly for driver reasons. If that rendertarget is 'simple'
enough (no custom shader, no custom texture coordinates, and nothing
else sharing the backing store), the rendertarget drawing will be
proxied unto the default EGL/... context of the video platform.

FossilOrigin-Name: a01f2f6bf9ccd27958f098f3a15c01c5cad768263a7b94f2135dac7da8a9e9d3