(shmif) handover_exec descriptor/env issue
This should fix two issues blocking handover allocation from working
properly, the first being a race where the parent could get to close the
descriptor before the child was finished creating, leading the server to
think that the connection point was dead.

The second was an overallocation in envv causing an extra garbage ptr
before the terminating NULL, preventing execve() from working.