(core/lua) tag-cb control for instant_transform
The tag image transform mechanism for getting callback notification on
completion of a transform stage in transformation chains masks all tags
by default in the case of an instant image transform on the backing
object.

This patch adds controls for adjusting if one or all transform tags
should be processed in such a case.