(core/lua/breaking) drop openctm support
We are a year or five passed the point of this ever getting a chance,
and 'the industry' seem to converge on glTF for mesh definitions, so
long term we do best in heading there as well.

As part of the general hardening of the core engine processes, this is a
decent chance to get rid of one of the bigger parsers still in use. The
problem is that the shmif- design is not there yet when it comes to
doing mesh- transfers, though aproto is prepared for it.

While afsrv_decode is reworked to handle model transfers, and the 3dbase
is adjusted to use a controlled decode- process for its transfers, the
workaround is to use the raw/table method of defining a mesh from lua.