(lua) allow loadstring with engine in debug-mode
We have historically been strict about being able to close down any
option for dynamic code generation from the .lua persective as opens up
for all kinds of obfuscation nightmares. The issue is that it also
closes some legitimate debug-use that makes it harder to develop with
the engine. To add a compromise, we enable loadstring if the engine
has been explicitly started with a higher debuglevel, i.e. arcan -g