(builtin) shutdown hookscript (hook/shutdown.lua)
This adds the command-line option for scheduling a shutdown call
after a certain amount of ticks. The default, 500 ticks, can be
regulated with an argument:

    arcan -H hook/shutdown.lua myappl shutdown=200

The purpose / use for this hook script is primarily for automation,
development and testing so that this doesn't have to be added to
every test case.