(builtin) add external_input hookscript
This adds a hook script that can be used to attach external input
providers. When the hook script is invocated, an extio_1 connection
point gets opened up. There are currently not that many input drivers,
as the normal approach is simply to use the device support provided by
the static / built-in platform layer.

Two simple examples can be found in tests/frameservers, drv_eyetrack
(tobii eye tracking support for the respective support libraries are
installed) and ioinject (random garbage).