(tools) add trayicon tool
This tools works with the 'extbtn' support in durden (and others if they
copy that model). It takes two icon arguments (passive, active) and a
chain-exec argument.

It connects and identifies as ICON as its primary type, then draws the
passive icon state. When clicked or otherwise triggered, it requests a
HANDOVER segment and inherits into a new process. While this process is
a live, the icon state is drawn as 'active'.