(egl-dri) switch default to no rtproxy
This feature has more complicated side-effects than is reasonable for it
to be a default for the time being. Thus we still pay the price for
full- re-composition for the time being, but hopefully it will be a
temporary thing while things change on the driver level.

FossilOrigin-Name: 426668dfb4d529f55bd7384ffc72a059c97fa41d6f7cd65177aacc03363356fd