(shmif) forward segment_token in NEWSEGMENT
This provides the segment_token as part of the newsegment. Normally this
is not needed as you retrieve it as part of the mapping of a segment,
but for HANDOVER connections, it might be needed to let the parent
reposition its child.

FossilOrigin-Name: 43f6b73039c67bec606d939b3e5b10fa26e314a72a5d55666d402cb96661039a