(egl-dri) fix to fade-flickers
This was interesting yet dumb. The effect only triggered for certain
platform configurations in cases where the last object in the chain
for a rendertarget had been fading to 0. The uniform and blend state
was not updated in the draw_display rt to egldisp. stage, thus the
alpha- from the last object would be retained.