(tools/aloadimage) allow hmd- l/r output
This is a work in progress as it needs actual support in safespaces
etc. to be testable. The general idea though is that you can prefix
the playlist items with a l: or r: and the respective eye would get
updated. Then, if we get a right- HMD segment we run two passes.

The main caveat with that would be that the image updates are not
necessarily synchronized. It is solvable with some locking games,
but lets first get it working partially.