(egl-dri) map platform_video_reset
This is a complicated one that will have to be iterated over in multiple
steps. Basically the feature deallocates all resources tied to any one
gpu and then rebuilds them on the same or a different gpu. Basically it
is a strong prerequisite for proper multi-GPU support, especially the
hotswap stage.