(tools) drop leddec
This exists in the aled- repository anyhow, and was not of much use in
its current shape. It is likely that this entire interface should be
dropped in favor of using shmif and encode- output with the leds simply
packed as a 2D texture. The 'clear all' and 'set all' case becomes a
little harder to translate, but bandwidth preserving encoders would
likely need copy+xor+rle before anyhow and detect those special cases.