(netproxy) restructure/refactor for crypto
The initial setup, key configuration and key-lookup interfaces were
moved around some to make it cleaner and less error prone to set up.
The first hello- packet structure was changed to use a salt with the
preshared auth-key.

Dropped the netpipe binary as it was mostly getting in the way, can
re-add with external-tunnel mode again when the other work on the main
proxy is finished.

FossilOrigin-Name: 1a675035f0ed53b6ed7cc0795030d61d7187ec4da30da9342449d3016dafcc50