(synch/agp/tui) |revert nonblock in tui +dirty
This should fix the 'requires screen update' for new synch, the exact
details are not entirely clear, only that the voluntary tearing in tui
turned out to often lead to signal without block but no refresh on synch.

The real fix for this comes with tui rework, with vpts updates on both
next-synch and last-acked. Then the SIGVID_BLOCK_NONE can be used for
skip-with-tear control like behavior.