(tui) bufferwnd- widget exit status controls
This modifies the bufferwnd widget to support all visual attributes to
be set by the lookup function, as well as adding an option for exit
control. This is useful in cases where the widget tui context may need
to allow the user to explicitly commit (shift + return) or cancel out
(escape).