(platform) add helper for event- producer thread
Recent updates to BSD seem to have added blocking like behavior to the
event processing. To try and troubleshoot that we need to try and deal
with threaded reading of the key input devices in order to defer any
buffering effects. Since this code was needed / generically useful
anyhow, the test code got cleaned up and added as a 'producer_thread'
interface as part of the platform.