(agp) add rendertarget_alloc hook
To be able to build rendertargets that use gbm- scanout compatible
formats, we need to intercept operations on mapped rendertargets.

This patch adds the hooks for the platform driver to intercept the
rendertarget allocator calls. This would allow it to initialize the
GL store with contents explicitly tied to a gbm scanout buffer.

The missing part would then, hopefully, only be to use this hook
and switch over to non-composite default scanout for the platform.

This should provide the ideal path for scanning out a rendertarget
mapped to a display, assuming that no post-processing shaders are
needed. In that case we switch back to composited output mode.

FossilOrigin-Name: 973a83cc26ff478f39989bd89e5e620f42803984881930b11130e3fa3247d1e7